chore: refactor for sanity
