+++ internal/scanner/stats.go
@@ -1,0 +1,22 @@
+package scanner
+
+import (
+	"os"
+	"strings"
+	"time"
+)
+
+type Stats struct {
+	FilesProcessed int
+	FilesSkipped   int
+	Duration       time.Duration
+	MetaFile       string
+	FileStats      map[string]int
+	ProcessedFiles []string
+	SkipPatterns   []string
+	Command        string
+}
+
+func BuildCommand() string {
+	return strings.Join(os.Args, " ")
+}
