+++ internal/tui/styles.go
@@ -1,0 +1,18 @@
+package tui
+
+import "github.com/charmbracelet/lipgloss"
+
+var (
+	focusedLabelStyle = lipgloss.NewStyle().
+				Foreground(lipgloss.Color("#00FF00")).
+				Bold(true)
+
+	focusedBorderStyle = lipgloss.NewStyle().
+				Border(lipgloss.NormalBorder()).
+				BorderForeground(lipgloss.Color("#00FF00")).
+				Padding(0, 1)
+
+	errorStyle = lipgloss.NewStyle().
+			Foreground(lipgloss.Color("#FF0000")).
+			Bold(true)
+)
